requirements.txt
paho-mqtt
pika